Move to 'pnc' module
